--- data_generation_pipeline/__init__.py
@@ -1,1 +1,0 @@
-# package marker
