+++ data_generation_pipeline/data-pipeline/config/model_config.yaml
@@ -1,0 +1,8 @@
+default_model: "Qwen/Qwen2.5-7B-Instruct"
+embed_model: "sentence-transformers/all-MiniLM-L6-v2"
+quantization:
+  load_in_4bit: true
+  bnb_4bit_use_double_quant: true
+  bnb_4bit_quant_type: "nf4"
+  bnb_4bit_compute_dtype: "float16"
+
