+++ data_generation_pipeline/data-pipeline/pyproject.toml
@@ -1,0 +1,13 @@
+[build-system]
+requires = ["setuptools", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "data-pipeline"
+version = "0.1.0"
+description = "Q/A generation pipeline for policy documents"
+authors = [
+  {name = "Saad Maqsood", email = "saadmaqsood1818@gmail.com"}
+]
+readme = "README.md"
+requires-python = ">=3.9"
