+++ data_generation_pipeline/data-pipeline/src/__init__.py
@@ -1,0 +1,2 @@
+__version__ = "0.1.0"
+
