+++ data_generation_pipeline/data-pipeline/src/chunker.py
@@ -1,0 +1,16 @@
+from typing import List
+
+def chunk_text(text: str, chunk_size: int=120, overlap: int=30) -> List[str]:
+    if not text:
+        return []
+    words = text.split()
+    if len(words) <= chunk_size:
+        return [" ".join(words)]
+    chunks = []
+    i = 0
+    while i < len(words):
+        chunk = words[i:i+chunk_size]
+        chunks.append(" ".join(chunk))
+        i += chunk_size - overlap
+    return chunks
+
