+++ data_generation_pipeline/data-pipeline/src/generator.py
@@ -1,0 +1,39 @@
+import time
+import torch
+
+def generate_with_retry(tokenizer, model, prompt: str, max_new_tokens: int=512,
+                        deterministic_temp: float=0.0, sampling_temp: float=0.4):
+    inputs = tokenizer(prompt, return_tensors="pt", truncation=True, max_length=tokenizer.model_max_length).to(model.device)
+    text = ""
+    try:
+        with torch.no_grad():
+            out = model.generate(
+                **inputs,
+                max_new_tokens=max_new_tokens,
+                temperature=deterministic_temp,
+                top_p=0.95,
+                do_sample=False,
+                pad_token_id=tokenizer.eos_token_id,
+                eos_token_id=tokenizer.eos_token_id,
+            )
+        text = tokenizer.decode(out[0][inputs["input_ids"].shape[1]:], skip_special_tokens=True).strip()
+    except Exception:
+        text = ""
+
+    if not text or len(text) < 10:
+        try:
+            with torch.no_grad():
+                out = model.generate(
+                    **inputs,
+                    max_new_tokens=max_new_tokens,
+                    temperature=sampling_temp,
+                    top_p=0.95,
+                    do_sample=True,
+                    pad_token_id=tokenizer.eos_token_id,
+                    eos_token_id=tokenizer.eos_token_id,
+                )
+            text = tokenizer.decode(out[0][inputs["input_ids"].shape[1]:], skip_special_tokens=True).strip()
+        except Exception:
+            text = ""
+    return text
+
